+++ modules/engine/src/cpp/ProgramOptions.cpp
@@ -102,6 +102,7 @@ ProgramOptions::ProgramOptions(wstringVector args, NELSON_ENGINE_MODE mode)
     _quietmode = false;
     _minimize = false;
     _ipc = false;
+    _withoutFileWatcher = false;
     _error.clear();
     _file.clear();
     _command.clear();
@@ -123,6 +124,7 @@ ProgramOptions::~ProgramOptions()
     _usermodules = false;
     _quietmode = false;
     _ipc = false;
+    _withoutFileWatcher = false;
     _minimize = false;
     _error.clear();
     _file.clear();
@@ -238,6 +240,8 @@ ProgramOptions::parse()
     Option nousermodulesOption(L"nousermodules", L"", _W("no user modules loaded"), false, false);
     Option minimizeOption(L"minimize", L"", _W("minimize main window (GUI only)"), false, false);
     Option noIpcOption(L"noipc", L"", _W("no ipc features"), false, false);
+    Option withoutFileWatcherOption(
+        L"withoutfilewatcher", L"", _W("without file watcher"), false, false);
     Option commandtoexecuteOption(L"execute", L"e", _W("command to execute"), false, true);
     Option filetoexecuteOption(L"file", L"f", _W("file to execute in an new process"), false, true);
     Option filetoexecuteIPCOption(
@@ -270,6 +274,7 @@ ProgramOptions::parse()
     _options = _options + timeoutOption.getFullDescription() + L"\n";
     _options = _options + openFilesOption.getFullDescription() + L"\n";
     _options = _options + loadFilesOption.getFullDescription() + L"\n";
+    _options = _options + withoutFileWatcherOption.getFullDescription() + L"\n";
 
     bRes = parseOption(helpOption, _ishelp);
     bRes = bRes && parseOption(versionOption, _isversion);
@@ -280,6 +285,7 @@ ProgramOptions::parse()
         bRes = bRes && parseOption(minimizeOption, _minimize);
     }
     bRes = bRes && parseOption(noIpcOption, _ipc);
+    bRes = bRes && parseOption(withoutFileWatcherOption, _withoutFileWatcher);
     bRes = bRes && parseOptionWithValues(openFilesOption, _filesToOpen);
     bRes = bRes && parseOptionWithValues(loadFilesOption, _filesToLoad);
     bool bFind = false;
@@ -411,6 +417,15 @@ ProgramOptions::haveNoIpc()
 {
     if (_isvalid) {
         return _ipc;
+    }
+    return false;
+}
+//=============================================================================
+bool
+ProgramOptions::haveWithoutFileWatcher()
+{
+    if (_isvalid) {
+        return _withoutFileWatcher;
     }
     return false;
 }